apps/app/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic[email]==2.1.1
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1